sensors: switch timestamps to CLOCK_BOOTIME
The Android CDD requires that timestamps 'must be synchronized with the
elapsedRealtimeNano clock'. elapsedRealtimeNano is CLOCK_BOOTIME.

All Kernel sensor drivers now report the timestamp in CLOCK_BOOTTIME
using the EV_SYN_TIME* event codes.

getTimestamp() is only used by the on-change sensors when they are first
turned on.

Increase the size of the input event  buffer to accommodate the
additional events for Accel, Gyro and Compass (3 axes, 2 timestamp and 1
sync).

[1] https://source.android.com/devices/sensors/hal-interface.html#sensors_event_t